backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
ImageHash
pytest
pytest-asyncio
httpx
